First draft of the configuration system - Ticket #32
This is the first draft of the configuration system to stop so many
commits being needed to fix the name in the repo. As per Bruce's
request.

This relies on the Java properties class and will need
uc_pircbotx.properties to be copied to the working directory in order to
work correctly. All options have defaults which are are the same as the
hard coded values being used.