Made all commands use the message interface
All commands now use the generic message interface rather than the
channel message event. This means it's now possible to code to a generic
interface and work with both private messages and with channel messages.

I have also added functionality which was not available before to the
models. It is now possible to send actions directly though the message
which is passed to the command.

This refixes issue #35