Fixed issue with onList throwing exception.
A few commits back we made the bot not as tightly coupled to message events so it would
work with private messages as well as channel messages (before pircbotx2). This means
the arguments in the methods need to be of type Message rather than the pircbotx class.

I missed this one so it didn't get updated... sorry :(.