Stuff to make the bot work as a daemon
This uses apache-daemon and jsvc to allow running of the bot as a unix daemon.
The code do this this mainly exists to make life easier for WebPigeon (me) so
he doesn't have to worry about restarting the bot.

At the moment it's using the default PID and user settings for jsvc, that's
not a good idea but it was for testing.

2 new scripts in the /src/main/scripts folder, one to generate a config and
one to invoke jsvc with the correct command line arguments. Both are really
basic and need much work to be production ready.