Fixed broken javadocs
Looks like I messed up a merge in 7a55eddfedd4b39ff8bdaa8ca83773ac04adde89

I'm not sure how that happened, but it's too far back in the history for
me to rebase fix (would cause a lot of problems). Instead this commit
fixes the problems introduced in that commit and allows the project's
javadoc to build again.