Fixed error duplication bug
Fixed error duplication which occoured when a module existed but the
requested command didn't exist and there was no default action defined
in the module.

Both the command processor and the irc handler would report the error,
now only the irc handler will report the error to the user. The error
will be logged by the command processor as a warning to its log.